Struct cosmo::collection::SpscConcurrentQueue

pub struct SpscConcurrentQueue<T> {
    // some fields omitted
}

A bounded queue allowing a single producer and a single consumer.

Methods

impl<T> SpscConcurrentQueue<T>

fn with_capacity(initial_capacity: usize) -> Arc<SpscConcurrentQueue<T>>

Creates a single producer single consumer queue with the specified capacity.

Examples

use cosmo::collection::{ConcurrentQueue, SpscConcurrentQueue};
let q = SpscConcurrentQueue::<u64>::with_capacity(1024);
assert_eq!(1024, q.capacity());

impl<T: Clone> SpscConcurrentQueue<T>

fn peek(&self) -> Option<T>

Tries to peek a value from the queue. The value needs to be Clone since the value in the queue can't be moved out.

If the queue is not empty, the method returns Some(v), where v is the value at the head of the queue. If the queue is empty, it returns None.

Examples

use cosmo::collection::{ConcurrentQueue, SpscConcurrentQueue};
let queue = SpscConcurrentQueue::<u64>::with_capacity(16);
match queue.peek() {
    Some(v) => println!("Peeked value {}", v),
    None => println!("Queue is empty")
}