Trait cosmo::collection::ConcurrentQueue

pub trait ConcurrentQueue<T> {
    fn offer(&self, val: T) -> Option<T>;
    fn poll(&self) -> Option<T>;
    fn capacity(&self) -> usize;
    fn size(&self) -> usize;

    fn is_empty(&self) -> bool { ... }
    fn is_full(&self) -> bool { ... }
}

Trait that concurrent, non-blocking queues implement.

Required Methods

fn offer(&self, val: T) -> Option<T>

Tries to put a value onto the queue.

If the queue is not full, the method returns None, signifying success. If the queue is full, it returns Some(v), where v is the original, specified value.

Examples

use cosmo::collection::{ConcurrentQueue, SpscConcurrentQueue};
let queue = SpscConcurrentQueue::<u64>::with_capacity(16);
match queue.offer(10) {
    Some(v) => println!("Queue is full"),
    None => println!("Value added to the queue")
}

fn poll(&self) -> Option<T>

Tries to remove a value from the queue.

If the queue is not empty, the method returns Some(v), effectively removing v from the queue. If the queue is empty, it returns None.

Examples

use cosmo::collection::{ConcurrentQueue, SpscConcurrentQueue};
let queue = SpscConcurrentQueue::<u64>::with_capacity(16);
match queue.poll() {
    Some(v) => println!("Removed item {}", v),
    None => println!("Queue is empty")
}

fn capacity(&self) -> usize

Returns the capacity of the queue.

Examples

use cosmo::collection::{ConcurrentQueue, SpscConcurrentQueue};
let queue = SpscConcurrentQueue::<u64>::with_capacity(16);
assert_eq!(16, queue.capacity());
queue.offer(10);
assert_eq!(16, queue.capacity());

fn size(&self) -> usize

Returns how many items are in the queue.

Examples

use cosmo::collection::{ConcurrentQueue, SpscConcurrentQueue};
let queue = SpscConcurrentQueue::<u64>::with_capacity(16);
assert_eq!(0, queue.size());
queue.offer(10);
assert_eq!(1, queue.size());

Provided Methods

fn is_empty(&self) -> bool

Tells whether the queue is empty or not.

Examples

use cosmo::collection::{ConcurrentQueue, SpscConcurrentQueue};
let queue = SpscConcurrentQueue::<u64>::with_capacity(16);
if queue.is_empty() {
    queue.offer(10);
}
assert_eq!(Some(10), queue.poll());

fn is_full(&self) -> bool

Tells whether the queue is full or not.

Examples

use cosmo::collection::{ConcurrentQueue, SpscConcurrentQueue};
let queue = SpscConcurrentQueue::<u64>::with_capacity(16);
if !queue.is_full() {
    queue.offer(10);
}
assert_eq!(Some(10), queue.poll());